requirements.txt
chromadb>=0.4.24
google-genai>=0.5.0
langchain>=0.1.10
langchain-google-genai>=0.0.9
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
pypdf>=4.0.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.1
